refactor(ui): rename file
